+++ src/Compute/Compute/VirtualMachine/Operation/NewAzureVMCommand.cs
@@ -255,6 +255,10 @@ namespace Microsoft.Azure.Commands.Compute
         [ValidateNotNullOrEmpty]
         public string UserAssignedIdentity { get; set; }
 
+        [Parameter(ParameterSetName = SimpleParameterSet, Mandatory = false, HelpMessage = "Use this to add the assign user specified identity (MSI) to the VM")]
+        [ValidateNotNullOrEmpty]
+        public string EncryptionIdentity { get; set; }
+
         [Parameter(Mandatory = false, HelpMessage = "Run cmdlet in the background")]
         public SwitchParameter AsJob { get; set; }
 
@@ -1351,22 +1355,37 @@ namespace Microsoft.Azure.Commands.Compute
         /// <returns>Returning the Identity generated form the cmdlet parameters "SystemAssignedIdentity" and "UserAssignedIdentity"</returns>
         private VirtualMachineIdentity GetVMIdentityFromArgs()
         {
-            var isUserAssignedEnabled = !string.IsNullOrWhiteSpace(UserAssignedIdentity);
-            return (SystemAssignedIdentity.IsPresent || isUserAssignedEnabled)
-                ? new VirtualMachineIdentity
-                {
-                    Type = !isUserAssignedEnabled ?
-                           CM.ResourceIdentityType.SystemAssigned :
-                           (SystemAssignedIdentity.IsPresent ? CM.ResourceIdentityType.SystemAssignedUserAssigned : CM.ResourceIdentityType.UserAssigned),
+            var isUserAssignedEnabled = !string.IsNullOrWhiteSpace(UserAssignedIdentity) || !string.IsNullOrWhiteSpace(EncryptionIdentity);
 
-                    UserAssignedIdentities = isUserAssignedEnabled
-                                             ? new Dictionary<string, UserAssignedIdentitiesValue>()
-                                             {
-                                                 { UserAssignedIdentity, new UserAssignedIdentitiesValue() }
-                                             }
-                                             : null,
+            if (SystemAssignedIdentity.IsPresent || isUserAssignedEnabled) {
+
+                Microsoft.Azure.Management.Compute.Models.ResourceIdentityType Type = CM.ResourceIdentityType.None;
+                Dictionary<string, UserAssignedIdentitiesValue> UserAssignedIdentities = new Dictionary<string, UserAssignedIdentitiesValue>();
+
+                if (!isUserAssignedEnabled) {
+                    Type = CM.ResourceIdentityType.SystemAssigned;
+                    UserAssignedIdentities = null;
                 }
-                : null;
+                else {
+                    if (SystemAssignedIdentity.IsPresent) {
+                        Type = CM.ResourceIdentityType.SystemAssignedUserAssigned;
+                    }
+                    else {
+                        Type = CM.ResourceIdentityType.UserAssigned;
+                    }
+
+                    if (!string.IsNullOrWhiteSpace(UserAssignedIdentity)) {
+                        UserAssignedIdentities.Add(UserAssignedIdentity,  new UserAssignedIdentitiesValue());
+                    }
+                    else if (!string.IsNullOrWhiteSpace(EncryptionIdentity)) {
+                        UserAssignedIdentities.Add(EncryptionIdentity,  new UserAssignedIdentitiesValue());
+                    }
+                }
+                return new VirtualMachineIdentity {Type = Type, UserAssignedIdentities = UserAssignedIdentities,};
+            }
+            else {
+                return null;
+            }
         }
 
         private string GetBginfoExtension()